PGDump: New option to truncate an existing table instead of recreating it
This commit adds a layer creation option "TRUNCATE_TABLE" that
performs a TRUNCATE at transaction start.

Like the already implemented DROP TABLE, this TRUNCATE cascades: it
truncates other tables that have foreign key dependency to the current
table before truncating the current table. This is a data integrity
risk, because the user may not suppose that other tables are going to
be altered by the command, but it allows OGR to recreate an entire
database by supplying the layers in the creation dependency order.

The TRUNCATE restarts identity: dependent sequences are reset to their
first value. This is a data consistency risk, because the updated rows
may not get the same serial identifier, and identifiers are reused for
possibly different data; but it creates a consistency with the
original data set: rows in PostgreSQL have the same fid as their GDAL
fid, and all rows before the first deleted row will keep their
original identifier.

Because of RESTART IDENTITY, this commit makes GDAL require PostgreSQL
>= 9.0 (first released in 2010).